+++ MvcStore/Models/Order.cs
@@ -39,7 +39,7 @@ namespace MvcStore.Models
         [RegularExpression(@"^[A-Za-z0-9](([_\.\-]?[a-zA-Z0-9]+)*)@([A-Za-z0-9]+)(([\.\-‌​]?[a-zA-Z0-9]+)*)\.([A-Za-z]{2,})$", ErrorMessage = "Email is not valid")]
         public string Email      { get; set; }
 
-        public float Total     { 
+        public double Total     { 
             get{return OrderCart.CartTotal;}
         }
         [NotMapped]
